Allow local development to target SQLite
When NODE_ENV is set to production, the project will still target a
postgres server, but if it isn't (like when it's 'just' run from a cli)
then it starts in development mode, targetting a local sqlite file.

+++ util/db/pre-migrations/010-create-schema.js
@@ -1,10 +1,23 @@
 'use strict';
 
-module.exports = {
-  up: (queryInterface) => {
-    return queryInterface.createSchema('traps');
-  },
-  down: (queryInterface) => {
-    return queryInterface.dropSchema('traps');
-  }
-};
+// The pre-migrations only make sense when running inside the production docker
+// environment. They are not required for the development SQLite DB.
+if (process.env.NODE_ENV === 'production') {
+  module.exports = {
+    up: (queryInterface) => {
+      return queryInterface.createSchema('traps');
+    },
+    down: (queryInterface) => {
+      return queryInterface.dropSchema('traps');
+    }
+  };
+} else {
+  module.exports = {
+    up: () => {
+      return Promise.resolve();
+    },
+    down: () => {
+      return Promise.resolve();
+    }
+  };
+}
